+++ tests/TestCase.php
@@ -1,10 +1,9 @@
 <?php
 /**
- * @package Mizmoz
- * @copyright Copyright 2016 Mizmoz Limited - Released under the MIT license
+ * @package mizmoz/validate
+ * @copyright Copyright 2018 Mizmoz Limited - Released under the MIT license
  * @see https://www.mizmoz.com/labs/validate
  */
-
 
 namespace Mizmoz\Validate\Tests;