+++ tests/Validator/IsReCaptchaTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * @package Mizmoz
- * @copyright Copyright 2016 Mizmoz Limited - Released under the MIT license
+ * @package mizmoz/validate
+ * @copyright Copyright 2018 Mizmoz Limited - Released under the MIT license
  * @see https://www.mizmoz.com/labs/validate
  */